--- internal/pipe/docker/docker_registry_test.go
@@ -1,49 +1,0 @@
-package docker
-
-import (
-	"flag"
-	"os"
-	"os/exec"
-	"testing"
-)
-
-var it = flag.Bool("it", false, "push images to docker hub")
-var registry = "localhost:5000/"
-var altRegistry = "localhost:5050/"
-
-func TestMain(m *testing.M) {
-	flag.Parse()
-	if *it {
-		registry = "docker.io/"
-	}
-	os.Exit(m.Run())
-}
-
-func start(t *testing.T) {
-	if *it {
-		return
-	}
-	if out, err := exec.Command(
-		"docker", "run", "-d", "-p", "5000:5000", "--name", "registry", "registry:2",
-	).CombinedOutput(); err != nil {
-		t.Log("failed to start docker registry", string(out), err)
-		t.FailNow()
-	}
-	if out, err := exec.Command(
-		"docker", "run", "-d", "-p", "5050:5000", "--name", "alt_registry", "registry:2",
-	).CombinedOutput(); err != nil {
-		t.Log("failed to start alternate docker registry", string(out), err)
-		t.FailNow()
-	}
-}
-
-func killAndRm(t *testing.T) {
-	if *it {
-		return
-	}
-	t.Log("killing registry")
-	_ = exec.Command("docker", "kill", "registry").Run()
-	_ = exec.Command("docker", "rm", "registry").Run()
-	_ = exec.Command("docker", "kill", "alt_registry").Run()
-	_ = exec.Command("docker", "rm", "alt_registry").Run()
-}
