+++ internal/pipe/docker/docker_test.go
@@ -1,6 +1,7 @@
 package docker
 
 import (
+	"flag"
 	"fmt"
 	"io/ioutil"
 	"os"
@@ -17,6 +18,47 @@ import (
 	"github.com/goreleaser/goreleaser/pkg/context"
 	"github.com/stretchr/testify/require"
 )
+
+var it = flag.Bool("it", false, "push images to docker hub")
+var registry = "localhost:5000/"
+var altRegistry = "localhost:5050/"
+
+func TestMain(m *testing.M) {
+	flag.Parse()
+	if *it {
+		registry = "docker.io/"
+	}
+	os.Exit(m.Run())
+}
+
+func start(t *testing.T) {
+	if *it {
+		return
+	}
+	if out, err := exec.Command(
+		"docker", "run", "-d", "-p", "5000:5000", "--name", "registry", "registry:2",
+	).CombinedOutput(); err != nil {
+		t.Log("failed to start docker registry", string(out), err)
+		t.FailNow()
+	}
+	if out, err := exec.Command(
+		"docker", "run", "-d", "-p", "5050:5000", "--name", "alt_registry", "registry:2",
+	).CombinedOutput(); err != nil {
+		t.Log("failed to start alternate docker registry", string(out), err)
+		t.FailNow()
+	}
+}
+
+func killAndRm(t *testing.T) {
+	if *it {
+		return
+	}
+	t.Log("killing registry")
+	_ = exec.Command("docker", "kill", "registry").Run()
+	_ = exec.Command("docker", "rm", "registry").Run()
+	_ = exec.Command("docker", "kill", "alt_registry").Run()
+	_ = exec.Command("docker", "rm", "alt_registry").Run()
+}
 
 // TODO: this test is too big... split in smaller tests? Mainly the manifest ones...
 func TestRunPipe(t *testing.T) {
